Jumping added. Notice the bugs added to this commit.
One of the bugs is the character going through the platforms.
This is because we are moving the character before checking for
collisions. In the frame before the collision, the character is no
where near the platform. But in the next frame, after adding the
velocities, the character will now be at the other side of the platform.

One solution to this bug is to check for collision before moving the
characters. In this case, we should extend the the collision box by the
velocity of the object.

Another solution is to limit the velocity to make sure that the
character will never go past the platform in one frame.